# cmd/portal/docs/swagger.yaml
definitions:
  dto.CreateNewCardReq:
    properties:
      card:
        properties:
          assignedTo:
            type: integer
          description:
            type: string
          duedate:
            type: integer
          title:
            type: string
        type: object
      columnID:
        type: integer
    type: object
...
  dto.CreateNewTeamReq:
    properties:
      teamName:
        type: string
    type: object
...
  dto.DeleteCardReq:
    properties:
      cardID:
        type: integer
    type: object
...
  dto.GetAllTeamByUserIDResp:
    properties:
      error:
        properties:
          code:
            type: integer
          message:
            type: string
        type: object
      teams:
        type: string
    type: object
...
  dto.GetTeamByIDResp:
    properties:
      error:
        properties:
          code:
            type: integer
          message:
            type: string
        type: object
      team:
        type: string
    type: object
...
  dto.ReorderCard:
    properties:
      aboveOfIdx:
        type: integer
      cardID:
        type: integer
      columnID:
        type: integer
    type: object
  dto.RespError:
    properties:
      error:
        properties:
          code:
            type: integer
          message:
            type: string
        type: object
    type: object
  dto.UpdateCardInfoReq:
    properties:
      card:
        properties:
          assignedTo:
            type: integer
          description:
            type: string
          duedate:
            type: integer
          id:
            type: integer
          title:
            type: string
        type: object
      columnID:
        type: integer
    type: object
...
  dto.UpdateTeamMemberReq:
    properties:
      memberID:
        type: integer
      op:
        type: string
    type: object
...
info:
  contact: {}
  license: {}
paths:
...
  /teams:
    get:
      description: Get all team by userID
      parameters:
      - description: jwt
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.GetAllTeamByUserIDResp'
      summary: Get all team by userID
      tags:
      - team
    post:
      description: Create new team
      parameters:
      - description: jwt
        in: header
        name: Authorization
        required: true
        type: string
      - description: team info
        in: body
        name: createReq
        required: true
        schema:
          $ref: '#/definitions/dto.CreateNewTeamReq'
      produces:
      - application/json
      responses:
        "200":
          description: Create OK
          schema:
            $ref: '#/definitions/dto.RespError'
      summary: Create new team
      tags:
      - team
  /teams/:teamID:
    delete:
      description: Delete team
      parameters:
      - description: jwt
        in: header
        name: Authorization
        required: true
        type: string
      - description: teamID
        in: path
        name: teamID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: delete OK
          schema:
            $ref: '#/definitions/dto.RespError'
      summary: Delete team
      tags:
      - team
    get:
      description: Get team by ID
      parameters:
      - description: jwt
        in: header
        name: Authorization
        required: true
        type: string
      - description: teamID
        in: path
        name: teamID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: team details
          schema:
            $ref: '#/definitions/dto.GetTeamByIDResp'
      summary: Get team by ID
      tags:
      - team
    put:
      description: Edit member by teamID and userID
      parameters:
      - description: jwt
        in: header
        name: Authorization
        required: true
        type: string
      - description: teamID
        in: path
        name: teamID
        required: true
        type: integer
      - description: Op =  add/remove
        in: body
        name: editReq
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateTeamMemberReq'
      produces:
      - application/json
      responses:
        "200":
          description: edit ok
          schema:
            $ref: '#/definitions/dto.RespError'
      summary: Edit member
      tags:
      - team
...
  /teams/:teamID/projects/:projectID/cards:
    delete:
      description: Delete card by ID
      parameters:
      - description: jwt
        in: header
        name: Authorization
        required: true
        type: string
      - description: teamID
        in: path
        name: teamID
        required: true
        type: integer
      - description: projectID
        in: path
        name: projectID
        required: true
        type: integer
      - description: delete request
        in: body
        name: deleteReq
        required: true
        schema:
          $ref: '#/definitions/dto.DeleteCardReq'
      produces:
      - application/json
      responses:
        "200":
          description: delete success
          schema:
            $ref: '#/definitions/dto.RespError'
      summary: Delete card by ID
      tags:
      - card
    patch:
      description: Update card info (No column ID)
      parameters:
      - description: jwt
        in: header
        name: Authorization
        required: true
        type: string
      - description: teamID
        in: path
        name: teamID
        required: true
        type: integer
      - description: projectID
        in: path
        name: projectID
        required: true
        type: integer
      - description: update info
        in: body
        name: updateReq
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateCardInfoReq'
      produces:
      - application/json
      responses:
        "200":
          description: Update OK
          schema:
            $ref: '#/definitions/dto.RespError'
      summary: Update card info
      tags:
      - card
    post:
      description: Create new card
      parameters:
      - description: jwt
        in: header
        name: Authorization
        required: true
        type: string
      - description: teamID
        in: path
        name: teamID
        required: true
        type: integer
      - description: projectID
        in: path
        name: projectID
        required: true
        type: integer
      - description: CreateReq
        in: body
        name: createReq
        required: true
        schema:
          $ref: '#/definitions/dto.CreateNewCardReq'
      produces:
      - application/json
      responses:
        "200":
          description: Create OK
          schema:
            $ref: '#/definitions/dto.RespError'
      summary: Create new card
      tags:
      - card
  /teams/:teamID/projects/:projectID/cards/reorder:
    post:
      description: Reorder card in one column
      parameters:
      - description: jwt
        in: header
        name: Authorization
        required: true
        type: string
      - description: teamID
        in: path
        name: teamID
        required: true
        type: integer
      - description: projectID
        in: path
        name: projectID
        required: true
        type: integer
      - description: ReorderReq
        in: body
        name: reorderReq
        schema:
          $ref: '#/definitions/dto.ReorderCard'
      produces:
      - application/json
      responses:
        "200":
          description: Reorder success
          schema:
            $ref: '#/definitions/dto.RespError'
      summary: Reorder card in one column
      tags:
      - card
...
swagger: "2.0"